nodistill.py file with 'add_noise' function
The added function 'add_noise' can be run as one of the arguments (--expt_type=) of the nodistill.py file. It adds Gaussian Noise to the labels generated by the expert model and trains the student on the newly created labels. It saves both the labels from the expert as well as the newly augmented labels.